requirements.txt
opencv-python>=4.0.0
Pillow>=8.0.0
torch>=1.7.0
ultralytics>=8.0.0
tqdm>=4.0.0
Flask>=2.3.0
Flask-CORS>=4.0.0
PyMuPDF>=1.23.0